requirements.txt
# Core ML
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
matplotlib==3.8.3
seaborn==0.13.2

# Deep Learning (optional)
# torch==2.2.1
# tensorflow==2.16.1

# MLOps Tools
mlflow==2.11.1
dvc==3.48.4
dvc-s3==3.2.0

# Data Engineering
apache-airflow==2.9.0
pyspark==3.5.1

# Model Serving
fastapi==0.110.1
uvicorn==0.29.0
flask==3.0.2

# Monitoring
evidently==0.4.16
great-expectations==0.18.12

# Experiment Tracking
optuna==3.6.1

# Feature Stores
feast==0.38.0

# Containerization / Cloud (CLI tools installed separately)
boto3==1.34.69
google-cloud-storage==2.16.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.4
requests==2.31.0
pytest==8.1.1
black==24.3.0
